rename to cli
