+++ src/main/java/convex/cli/Constants.java
@@ -1,4 +1,4 @@
-package convex.cli2;
+package convex.cli;
 
 
 /**
