+++ src/main/java/convex/cli/Peer.java
@@ -1,4 +1,4 @@
-package convex.cli2;
+package convex.cli;
 
 import java.util.HashMap;
 import java.util.Map;
